refactor: extract ExportDialog component and fix client name in export filename
